backend/pyproject.toml
[project]
name = "spotter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-health-check>=3.24.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "celery>=5.5.1",
    "gunicorn>=25.0.3",
    "psycopg2>=2.9.11",
    "python-decouple>=3.8",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "django-debug-toolbar>=6.1.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-env>=1.1.3",
    "ruff>=0.15.0",
]
